Move listing section as a shared component.
