+++ client/a8c-for-agencies/sections/marketplace/products-overview/product-listing/index.tsx
@@ -13,12 +13,12 @@ import { useDispatch } from 'calypso/state';
 import { recordTracksEvent } from 'calypso/state/analytics/actions';
 import FilterSearch from '../../../../components/filter-search';
 import { ShoppingCartContext } from '../../context';
+import ListingSection from '../../listing-section';
 import MultiProductCard from '../multi-product-card';
 import ProductCard from '../product-card';
 import useProductAndPlans from './hooks/use-product-and-plans';
 import { getSupportedBundleSizes, useProductBundleSize } from './hooks/use-product-bundle-size';
 import useSubmitForm from './hooks/use-submit-form';
-import ProductListingSection from './sections';
 import VolumePriceSelector from './volume-price-selector';
 import type { ShoppingCartItem } from '../../types';
 import type { SiteDetails } from '@automattic/data-stores';
@@ -294,7 +294,7 @@ export default function ProductListing( { selectedSite, suggestedProduct }: Prod
 			</div>
 
 			{ plans.length > 0 && (
-				<ProductListingSection
+				<ListingSection
 					icon={ <JetpackLogo size={ 26 } /> }
 					title={ translate( 'Jetpack Plans' ) }
 					description={ translate(
@@ -303,11 +303,11 @@ export default function ProductListing( { selectedSite, suggestedProduct }: Prod
 					isTwoColumns
 				>
 					{ getProductCards( plans ) }
-				</ProductListingSection>
+				</ListingSection>
 			) }
 
 			{ products.length > 0 && (
-				<ProductListingSection
+				<ListingSection
 					icon={ <JetpackLogo size={ 26 } /> }
 					title={ translate( 'Jetpack Products' ) }
 					description={ translate(
@@ -315,11 +315,11 @@ export default function ProductListing( { selectedSite, suggestedProduct }: Prod
 					) }
 				>
 					{ getProductCards( products ) }
-				</ProductListingSection>
+				</ListingSection>
 			) }
 
 			{ wooExtensions.length > 0 && (
-				<ProductListingSection
+				<ListingSection
 					icon={ <WooLogo width={ 45 } height={ 28 } /> }
 					title={ translate( 'WooCommerce Extensions' ) }
 					description={ translate(
@@ -327,11 +327,11 @@ export default function ProductListing( { selectedSite, suggestedProduct }: Prod
 					) }
 				>
 					{ getProductCards( wooExtensions ) }
-				</ProductListingSection>
+				</ListingSection>
 			) }
 
 			{ backupAddons.length > 0 && (
-				<ProductListingSection
+				<ListingSection
 					icon={ <JetpackLogo size={ 26 } /> }
 					title={ translate( 'Jetpack VaultPress Backup Add-ons' ) }
 					description={ translate(
@@ -339,7 +339,7 @@ export default function ProductListing( { selectedSite, suggestedProduct }: Prod
 					) }
 				>
 					{ getProductCards( backupAddons ) }
-				</ProductListingSection>
+				</ListingSection>
 			) }
 		</div>
 	);
